refactor common banner functionality into new file; remove info roundel
